curlpp: 0.7.3 -> 0.8.1
Upgrade curlpp from 0.7.3 to 0.8.1. The main difference between 0.7.x
and 0.8.0 is that it replaces autotools with cmake, see
https://github.com/jpbarrette/curlpp/releases/tag/v0.8.0

The homepage of curlpp on googlecode is obsoleted, so update it and use
soure code repo on github.

Remove dependency boost which is dropped by upstream. And remove extra
CXXFLAGS which has been fixed by upstream.

Signed-off-by: Kai Kang <kai.kang@windriver.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>